Add files via upload
In this project, we make a voting machine in which we have 4 different option to vote each user vote one time and after all the user gives their vote we will conclude the result and store in our file and we have a full-fledged file in which we store a list of winner party for every year. With this Voting System we can give an user a safe and good Voting environment without any scam in count of vote.